refactor + remove hello message
